auth/authentication/models.py
import re

from django.contrib.auth.models import AbstractUser, BaseUserManager
from django.core.exceptions import ValidationError
from django.db import models
from phonenumber_field.modelfields import PhoneNumberField
from django.utils.translation import gettext_lazy as _


class CustomUserManager(BaseUserManager):
    def create_user(self, email, password=None, **extra_fields):
        if not email:
            raise ValueError(_('The Email must be set'))
        email = self.normalize_email(email)
        user = self.model(email=email, **extra_fields)
        user.set_password(password)
        user.full_clean()
        user.save(using=self._db)
        return user

    def create_superuser(self, email, password, **extra_fields):
        extra_fields.setdefault('is_staff', True)
        extra_fields.setdefault('is_superuser', True)
        return self.create_user(email, password, **extra_fields)


class CustomUser(AbstractUser):
    username = None
    email = models.EmailField(_('email address'), unique=True)
    is_property_owner = models.BooleanField(default=False)
    is_admin = models.BooleanField(default=False)

    USERNAME_FIELD = 'email'
    REQUIRED_FIELDS = []

    objects = CustomUserManager()

    def __str__(self):
        return self.email

    @property
    def is_admin_user(self):
        return self.is_admin

    @property
    def is_property_owner_user(self):
        return self.is_property_owner

    def clean(self):
        super().clean()
        if not re.match("^[a-zA-Z]*$", self.first_name):
            raise ValidationError(_('First name must contain only letters.'))
        if not re.match("^[a-zA-Z]*$", self.last_name):
            raise ValidationError(_('Last name must contain only letters.'))


class UserProfile(models.Model):
    user = models.OneToOneField(CustomUser, on_delete=models.CASCADE, related_name='profile')
    photo = models.ImageField(upload_to='photos', null=True, blank=True, default='Default.png')
    phone_number = PhoneNumberField(unique=True, null=True, blank=True)
    additional_info = models.TextField(_('additional information'), blank=True)
    created_at = models.DateTimeField(auto_now_add=True)
    modified_at = models.DateTimeField(auto_now=True)

    def __str__(self):
        return f"{self.user.email}'s profile"
