auth/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.24
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
